+++ CHANGELOG.md
@@ -55,7 +55,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ### Changed
 - run_program:  Added repset_hosts to mongo_class.RepSet instance call.
-- chk_mem_rep_lag:  Changed dictionary to CamelCase.
+- chk_mem_rep_lag:  Changed dictionary to PascalCase.
 - chk_mem_rep_lag:  Refactored to handle only the JSON format and convert to standard out format if requested.
 - Documentation updates.