+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [3.6.0] - 2021-09-07
 ### Changed
+- config/mongo.py.TEMPLATE:  Added SSL connection entries.
 - Removed unnesscary \*\*kwargs from argument lists.
 - Documentation updates.